+++ src/classes/babies/ShopkeeperBaby.ts
@@ -3,16 +3,11 @@ import {
   PickupPrice,
   RoomType,
 } from "isaac-typescript-definitions";
-import {
-  Callback,
-  asNumber,
-  inRoomType,
-  levelHasRoomType,
-} from "isaacscript-common";
+import { Callback, levelHasRoomType } from "isaacscript-common";
 import { Baby } from "../Baby";
 
 /**
- * Free shop items.
+ * Items that cost coins are free.
  *
  * We have to use both the `POST_PICKUP_INIT` and the `POST_PICKUP_UPDATE` callbacks because the
  * price of rerolled items is `PickupPrice.NULL` in `POST_PICKUP_INIT`.
@@ -24,21 +19,15 @@ export class ShopkeeperBaby extends Baby {
 
   @Callback(ModCallback.POST_PICKUP_INIT)
   postPickupInit(pickup: EntityPickup): void {
-    this.makePickupFree(pickup);
+    if (pickup.Price > 0) {
+      pickup.Price = PickupPrice.FREE;
+      pickup.AutoUpdatePrice = false;
+    }
   }
 
   @Callback(ModCallback.POST_PICKUP_UPDATE)
   postPickupUpdate(pickup: EntityPickup): void {
-    this.makePickupFree(pickup);
-  }
-
-  makePickupFree(pickup: EntityPickup): void {
-    if (
-      pickup.Price !== asNumber(PickupPrice.NULL) &&
-      pickup.Price !== asNumber(PickupPrice.YOUR_SOUL) &&
-      pickup.Price !== asNumber(PickupPrice.FREE) &&
-      inRoomType(RoomType.SHOP, RoomType.ERROR)
-    ) {
+    if (pickup.Price > 0) {
       pickup.Price = PickupPrice.FREE;
       pickup.AutoUpdatePrice = false;
     }