+++ docs/history.md
@@ -6,6 +6,7 @@
 ### _v1.9.1_ - September 11th, 2023
 
 - Changed the following babies:
+  - Shopkeeper Baby (#215) - Items that cost coins are free
   - Lucky Baby (#563) - Starts with The Stairway (improved)
 - Fixed the bug where charged items would be incorrectly triggered when going to a new floor. (Thanks Adrayon)